feat: Use Enformer instead of EnformerIA3
